requirements.txt
openstacksdk==4.4.0
